pipeline, block thread only during task duration
Right now pipeline would block on channel consumption
that means that even when the pipeline has nothing to do we block/use
threads for nothing. This patch makes it so that we call (async/thread
f) only for the duration of the task and let the chan consumption
happen in a go block.